Refactor: extract subcomponents from AssessmentResult and ClinicalCopilot to address prop drilling (#1448)
