+++ client/src/components/AssessmentResult.tsx
@@ -18,6 +18,8 @@ import { BiomarkerAlerts } from "./BiomarkerAlerts";
 import { ClinicalAttentionNavigator } from "./ClinicalAttentionNavigator";
 import { ClinicalCopilot } from "./ClinicalCopilot";
 import { ClinicalNoteViewer } from "./ClinicalNoteViewer";
+import { ExplainabilityPanel } from "./assessment/ExplainabilityPanel";
+import { PathToImprovement } from "./assessment/PathToImprovement";
 import { Tooltip as UiTooltip, TooltipTrigger, TooltipContent } from "@/components/ui/tooltip";
 import { Textarea } from "@/components/ui/textarea";
 import { useTranslation } from "react-i18next";
@@ -32,7 +34,7 @@ interface RiskFactor {
   description: string;
 }
 
-interface FactorBreakdown extends RiskFactor {
+export interface FactorBreakdown extends RiskFactor {
   strength: number;
   plainReason: string;
 }
@@ -651,145 +653,3 @@ export function AssessmentResult({ assessment }: AssessmentResultProps) {
   );
 }
 
-function ExplainabilityPanel({
-  factors,
-  increasedRiskFactors,
-  reducedRiskFactors,
-}: {
-  factors: FactorBreakdown[];
-  increasedRiskFactors: FactorBreakdown[];
-  reducedRiskFactors: FactorBreakdown[];
-}) {
-  const { t } = useTranslation();
-  if (factors.length === 0) {
-    return null;
-  }
-
-  return (
-    <div className="bg-card border border-border rounded-xl p-5 sm:p-6 shadow-sm">
-      <div className="flex flex-col gap-2 sm:flex-row sm:items-start sm:justify-between mb-5">
-        <div>
-          <h3 className="font-bold text-lg flex items-center gap-2">
-            <Info className="w-5 h-5 text-primary" /> {t("patientResult.explainability")}
-          </h3>
-          <p className="text-sm text-muted-foreground mt-1">
-            {t("patientResult.explainabilityDesc")}
-          </p>
-        </div>
-        <div className="grid grid-cols-2 gap-2 text-xs font-semibold">
-          <span className="inline-flex items-center gap-1 rounded-full border border-red-200 bg-red-50 px-3 py-1 text-red-700">
-            <TrendingUp className="w-3.5 h-3.5" />
-            {increasedRiskFactors.length} {t("patientResult.raised")}
-          </span>
-          <span className="inline-flex items-center gap-1 rounded-full border border-green-200 bg-green-50 px-3 py-1 text-green-700">
-            <TrendingDown className="w-3.5 h-3.5" />
-            {reducedRiskFactors.length} {t("patientResult.reduced")}
-          </span>
-        </div>
-      </div>
-
-      <div className="space-y-4">
-        {factors.map((factor) => {
-          const increasesRisk = factor.impact === "positive";
-          return (
-            <div
-              key={`${factor.name}-${factor.impact}`}
-              className="rounded-lg border border-border/70 bg-muted/20 p-4"
-            >
-              <div className="flex flex-col gap-2 sm:flex-row sm:items-center sm:justify-between">
-                <div className="min-w-0">
-                  <p className="font-semibold text-foreground">{factor.name}</p>
-                  <p className="text-sm text-muted-foreground mt-1">{factor.plainReason}</p>
-                </div>
-                <span
-                  className={cn(
-                    "inline-flex shrink-0 items-center gap-1 rounded-full px-3 py-1 text-xs font-bold",
-                    increasesRisk
-                      ? "bg-red-50 text-red-700 border border-red-200"
-                      : "bg-green-50 text-green-700 border border-green-200"
-                  )}
-                >
-                  {increasesRisk ? (
-                    <TrendingUp className="w-3.5 h-3.5" />
-                  ) : (
-                    <TrendingDown className="w-3.5 h-3.5" />
-                  )}
-                  {increasesRisk ? t("patientResult.increasesRisk") : t("patientResult.reducesRisk")}
-                </span>
-              </div>
-
-              <div className="mt-4">
-                <div className="flex items-center justify-between text-xs font-medium text-muted-foreground mb-1.5">
-                  <span>{t("patientResult.relativeContribution")}</span>
-                  <span>{factor.strength}%</span>
-                </div>
-                <div className="h-2.5 rounded-full bg-muted overflow-hidden">
-                  <div
-                    className={cn("h-full rounded-full w-[var(--factor-strength)]", increasesRisk ? "bg-red-500" : "bg-green-500")}
-                    style={{ '--factor-strength': `${factor.strength}%` } as React.CSSProperties}
-                  />
-                </div>
-              </div>
-            </div>
-          );
-        })}
-      </div>
-    </div>
-  );
-}
-
-function PathToImprovement({ assessment }: { assessment: AssessmentResponse }) {
-  const { t } = useTranslation();
-  const { mutate, data, isPending } = useWhatIfAuto();
-
-  useEffect(() => {
-    if (!assessment) return;
-    mutate({
-      patientName: assessment.patientName ?? "Unknown",
-      gender: (assessment.gender as "Male" | "Female") || "Male",
-      age: assessment.age ?? 0,
-      hypertension: assessment.hypertension ?? false,
-      heartDisease: assessment.heartDisease ?? false,
-      smokingHistory: (assessment.smokingHistory as "current" | "never" | "No Info" | "former") || "No Info",
-      bmi: assessment.bmi ?? 25,
-      hba1cLevel: assessment.hba1cLevel ?? 5.5,
-      bloodGlucoseLevel: assessment.bloodGlucoseLevel ?? 100,
-    });
-  }, [assessment, mutate]);
-
-  if (isPending || !data) {
-    return (
-      <div className="bg-card border border-border rounded-xl p-5 shadow-sm animate-pulse flex items-center justify-center min-h-[100px]">
-        <Loader2 className="w-6 h-6 animate-spin text-primary mr-2" />
-        <span className="text-muted-foreground text-sm">{t("patientResult.analyzing")}</span>
-      </div>
-    );
-  }
-
-  const recommendations = (data as any)?.recommendations;
-  if (!recommendations || recommendations.length === 0) {
-    return null;
-  }
-
-  return (
-    <div className="bg-gradient-to-br from-green-50 to-emerald-100 border border-green-200 rounded-xl p-6 shadow-sm relative overflow-hidden">
-      <div className="absolute top-0 right-0 p-4 opacity-10">
-        <TrendingDown className="w-24 h-24 text-green-700" />
-      </div>
-      <h3 className="font-bold text-lg mb-4 flex items-center gap-2 text-green-900 relative z-10">
-        <TrendingDown className="w-5 h-5" /> {t("patientResult.pathToImprovement")}
-      </h3>
-      <div className="space-y-4 relative z-10">
-        {recommendations.map((rec: any, idx: number) => (
-          <div key={idx} className="bg-white/80 backdrop-blur-sm p-4 rounded-lg border border-green-200/50 flex gap-3 shadow-sm">
-            <CheckCircle2 className="w-5 h-5 text-green-600 shrink-0 mt-0.5" />
-            <div>
-              <p className="font-semibold text-green-900">{rec.action}</p>
-              <p className="text-green-800/80 text-sm mt-1">{rec.message}</p>
-            </div>
-          </div>
-        ))}
-      </div>
-    </div>
-  );
-}
